+++ person.h
@@ -1,0 +1,8 @@
+Name name;
+Date date_of_birth;
+int age;
+char gender[30];
+int phone_number;
+char email[50];
+Role role;
+Address address;
