+++ schoolRoster.c
@@ -1,75 +1,6 @@
 #include <stdio.h>
 #include <string.h>
-
-typedef enum {
-    TEACHER,
-    STAFF,
-    STUDENT
-} Role;
-
-typedef enum {
-    MATH,
-    SCIENCE,
-    ENGLISH,
-    READING,
-    WRITING
-} Department;
-
-struct TEACHER {
-    int salary;
-    Department department;
-};
-
-struct STUDENT {
-    char grade;
-};
-
-typedef enum {
-    SCHOOL_COUNSELOR,
-    DEAN,
-    PRINCIPAL,
-    ASSISTANT_PRINCIPAL,
-    JANITOR,
-    LIBRARIAN,
-    ATHLETICS_DIRECTOR
-} Position;
-
-struct STAFF {
-    int salary;
-    Position position;
-};
-
-typedef struct {
-    char street[50];
-    char apartment[10];
-    char city[50];
-    char state[3]; // 2-Letter state code and null terminator
-    char zip[10];
-} Address;
-
-typedef struct {
-    char first_name[25];
-    char middle_name[25];
-    char last_name[25];
-} Name;
-
-typedef struct {
-    int month;
-    int day;
-    int year;
-} Date;
-
-struct Person {
-    Name name;
-    Date date_of_birth;
-    int age;
-    char gender[30];
-    int phone_number;
-    char email[50];
-    char password[30];
-    Role role;
-    Address address;
-};
+#include "dao.h"
 
 /* Write to file */
